voice-agent/requirements.txt
fastapi
uvicorn
twilio
openai
python-dotenv
pymongo
requests
python-multipart
pyngrok
